requirements.txt
motor>=3.1.1
aiohttp>=3.8.0
python-dotenv>=1.0
sentence-transformers>=2.2.2
torch>=2.0.0
faiss-cpu>=1.7.2
concurrent-log-handler>=0.9.13
pandas>=1.5.0
requests>=2.28.0
